Add email cell for sender side
